fixed time

+++ RaphEngine/include/RaphEngine.h
@@ -16,8 +16,8 @@
 
 class RAPHENGINE_API Time {
 public:
-	static float deltaTime;
-	static long GetTime();
+	static double deltaTime;
+	static double GetTime();
 };
 
 class RaphEngine
